pyproject.toml
[project]
name = "adk-dataflow"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk[a2a]>=1.10.0",
    "uvicorn>=0.35.0",
]
